correction of mistakes

+++ classinf.h
@@ -7,7 +7,7 @@ class ClassInf
 { // работа с полями: факт существования, размер и имя файла
 private:
     bool f_exist; // статус: true or false
-    int f_size;
+    qint64 f_size;
     QString f_name;
 
 public:
